Inline point actions (list)
