+++ src/components/_shared/SelectedData/_parts/SelectedDataList/SelectedDataItem/SelectedDataItem.module.scss
@@ -1,4 +1,6 @@
 .selectedDataItem {
+  display: flex;
+  align-items: center;
   border: 1px solid rgba(#e5e7fa, 0.5);
   font-size: 10px;
   border-radius: 6px;
@@ -12,6 +14,7 @@
 
   > button {
     display: flex;
+    justify-content: space-between;
     align-items: center;
     width: 100%;
     padding: 7.5px;
@@ -55,4 +58,9 @@
   top: 50%;
   transform: translateY(-50%);
   right: 10px;
+}
+
+.SelectedDataItemActions {
+  display: flex;
+  padding-right: 10px;
 }
